28/12/2024: rename packages (del "_" from the name)
