+++ src/test/java/user/UserDeleteTest.java
@@ -1,7 +1,7 @@
 package user;
 
-import api.user_steps.UserStepsApi;
-import api.user_steps.UserStepsChecks;
+import api.usersteps.UserStepsApi;
+import api.usersteps.UserStepsChecks;
 import io.qameta.allure.Epic;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.ValidatableResponse;
@@ -9,7 +9,7 @@ import org.apache.commons.lang3.StringUtils;
 import org.junit.Before;
 import org.junit.Test;
 import pojo.User;
-import random_data.UserGeneratorData;
+import randomdata.UserGeneratorData;
 
 @Epic("Delete user")
 public class UserDeleteTest {
